Move class is::core::SystemHandle to is::SystemHandle namespace
