+++ packages/core/src/systemhandle/RegisterSystem.cpp
@@ -16,13 +16,12 @@
  *
  */
 
-#include <is/core/systemhandle/RegisterSystem.hpp>
+#include <is/systemhandle/RegisterSystem.hpp>
 
 #include <iostream>
 
 namespace eprosima {
 namespace is {
-namespace core {
 namespace internal {
 
 //==============================================================================
@@ -130,6 +129,5 @@ void register_system_handle_factory(
 }
 
 } //  namespace detail
-} //  namespace core
 } //  namespace is
 } //  namespace eprosima
